requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyGithub==1.25.0
Werkzeug==0.9.6
boto==2.29.1
gunicorn==19.0.0
itsdangerous==0.24
requests==2.3.0
wsgiref==0.1.2
